@@ -6,6 +6,6 @@ export default async function handle(_req, res) {
     return res.json(events);
   } catch (e) {
     console.log(e);
-    res.status(500).json({ error: "Internal server error" });
+    return res.status(500).json({ error: "Internal server error" });
   }
 }
