@@ -1,9 +1,5 @@
-import {
-    fetchChecmicumFoodlist
-} from "../../../services/unicafeFoodListService";
-  
+import { fetchChecmicumFoodlist } from "../../../services/unicafeFoodListService";
+
 export default function handle(req, res) {
-    fetchChecmicumFoodlist().then(foodlist => {
-        res.json(foodlist);
-    });
+  return fetchChecmicumFoodlist().then(foodlist => res.json(foodlist));
 }
