@@ -1,10 +1,5 @@
-import {
-    fetchExactumFoodlist,
-} from "../../../services/unicafeFoodListService";
-  
+import { fetchExactumFoodlist } from "../../../services/unicafeFoodListService";
+
 export default function handle(req, res) {
-    fetchExactumFoodlist().then(foodlist => {
-        res.json(foodlist);
-    });
+  return fetchExactumFoodlist().then(foodlist => res.json(foodlist));
 }
-  
